@@ -1,6 +1,6 @@
 package com.example.demo.entities;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.util.Date;
 
 
